requirements.txt
agate==1.13.0
agate-dbf==0.2.3
agate-excel==0.4.1
agate-sql==0.7.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.16.2
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
billiard==4.2.1
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
category_encoders==2.8.1
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
configobj==5.0.9
contourpy==1.3.2
cryptography==45.0.4
csvkit==2.1.0
cycler==0.12.1
dask==2025.5.1
databricks-sdk==0.57.0
datasets==3.6.0
dbfread==2.0.7
debugpy==1.8.14
decorator==5.2.1
dictdiffer==0.9.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
Django==5.2.3
docker==7.1.0
dpath==2.2.0
dulwich==0.22.8
durationpy==0.10
dvc==3.60.1
dvc-data==3.16.10
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-studio-client==0.21.0
dvc-task==0.40.2
dvclive==0.9.0
entrypoints==0.4
et_xmlfile==2.0.0
evaluate==0.4.4
executing==2.2.0
fastapi==0.115.14
filelock==3.18.0
Flask==3.1.1
flatbuffers==25.2.10
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.58.0
frozenlist==1.7.0
fsspec==2025.3.0
funcy==2.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.3
google-play-scraper==1.2.7
googleapis-common-protos==1.70.0
googletrans==4.0.0rc1
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.3
grpcio==1.73.1
gto==1.7.2
h11==0.16.0
h2==3.2.0
hf-xet==1.1.5
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.0
humanfriendly==10.0
hydra-core==1.3.2
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
kombu==5.5.4
kubernetes==33.1.0
leather==0.4.0
llvmlite==0.44.0
locket==1.0.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mlflow==3.1.1
mlflow-skinny==3.1.1
mmh3==5.1.0
monotonic-binning==0.0.1
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
nanotime==0.5.2
narwhals==1.40.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
numba==0.61.2
numpy==2.2.5
oauthlib==3.3.1
olefile==0.47
omegaconf==2.3.0
onnxruntime==1.22.0
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parsedatetime==2.6
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
patsy==1.0.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
posthog==5.4.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pyaes==1.6.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pydot==4.0.0
pygit2==1.18.0
Pygments==2.19.1
pygtrie==2.5.0
pyodbc==5.2.0
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==8.4.1
python-benedict==0.34.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-fsutil==0.15.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
scmrepo==3.3.11
scorecardpy==0.1.9.7
seaborn==0.13.2
semver==3.0.4
setuptools==80.9.0
shap==0.48.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
sonopy==0.1.2
SQLAlchemy==2.0.41
sqlparse==0.5.3
sqltrie==0.11.2
stack-data==0.6.3
starlette==0.46.2
statsmodels==0.14.4
streamlit==1.45.1
sympy==1.14.0
tabulate==0.9.0
Telethon==1.40.0
tenacity==9.1.2
text-unidecode==1.3
textblob==0.19.0
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.3
toolz==1.0.0
torch==2.7.1
tornado==6.5
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.35.0
vader==0.0.3
vine==5.1.0
voluptuous==0.15.2
waitress==3.0.2
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wordcloud==1.9.4
xgboost==3.0.2
xlrd==2.0.2
xxhash==3.5.0
yarl==1.20.1
zc.lockfile==3.0.post1
zipp==3.23.0

streamlit
sentence-transformers
transformers
faiss-cpu
torch
accelerate
bitsandbytes
huggingface_hub